feat: 작품정보 조회 API 추가
