+++ app/src/main/java/com/teamwss/websoso/data/repository/NovelRepository.kt
@@ -2,6 +2,7 @@ package com.teamwss.websoso.data.repository
 
 import com.teamwss.websoso.data.mapper.toData
 import com.teamwss.websoso.data.model.NovelDetailEntity
+import com.teamwss.websoso.data.model.NovelInfoEntity
 import com.teamwss.websoso.data.remote.api.NovelApi
 import javax.inject.Inject
 
@@ -19,5 +20,9 @@ class NovelRepository @Inject constructor(
 
     suspend fun deleteUserInterest(novelId: Long) {
         novelApi.deleteUserInterest(novelId)
+    }
+
+    suspend fun getNovelInfo(novelId: Long): NovelInfoEntity {
+        return novelApi.getNovelInfo(novelId).toData()
     }
 }
