+++ app/src/main/java/com/teamwss/websoso/ui/mapper/NovelMapper.kt
@@ -1,7 +1,13 @@
 package com.teamwss.websoso.ui.mapper
 
+import com.teamwss.websoso.data.model.NovelInfoEntity
 import com.teamwss.websoso.domain.model.NovelDetail
 import com.teamwss.websoso.ui.novelDetail.model.NovelDetailModel
+import com.teamwss.websoso.ui.novelInfo.model.KeywordModel
+import com.teamwss.websoso.ui.novelInfo.model.NovelInfoUiModel
+import com.teamwss.websoso.ui.novelInfo.model.PlatformModel
+import com.teamwss.websoso.ui.novelInfo.model.ReviewCountModel
+import com.teamwss.websoso.ui.novelInfo.model.UnifiedReviewCountModel
 import com.teamwss.websoso.ui.novelRating.model.ReadStatus
 
 fun NovelDetail.toUi(): NovelDetailModel {
@@ -39,3 +45,27 @@ fun NovelDetail.toUi(): NovelDetailModel {
         ),
     )
 }
+
+fun NovelInfoEntity.toUi() = NovelInfoUiModel(
+    novelDescription = novelDescription,
+    attractivePoints = attractivePoints,
+    unifiedReviewCount = reviewCount.toUi(),
+)
+
+fun NovelInfoEntity.PlatformEntity.toUi() = PlatformModel(
+    platformName = platformName,
+    platformImage = platformImage,
+    platformUrl = platformUrl,
+    isVisible = true,
+)
+
+fun NovelInfoEntity.KeywordEntity.toUi() = KeywordModel(
+    keywordName = keywordName,
+    keywordCount = keywordCount,
+)
+
+fun NovelInfoEntity.ReviewCountEntity.toUi() = UnifiedReviewCountModel(
+    watchingCount = ReviewCountModel(ReadStatus.WATCHING, watchingCount),
+    watchedCount = ReviewCountModel(ReadStatus.WATCHED, watchedCount),
+    quitCount = ReviewCountModel(ReadStatus.QUIT, quitCount),
+)
