+++ app/src/main/res/layout/activity_novel_detail.xml
@@ -97,6 +97,7 @@
                             android:id="@+id/iv_novel_detail_cover_genre"
                             android:layout_width="72dp"
                             android:layout_height="72dp"
+                            app:isVectorImage="@{true}"
                             android:scaleType="fitXY"
                             app:layout_constraintBottom_toBottomOf="@id/iv_novel_detail_cover"
                             app:layout_constraintEnd_toEndOf="@id/iv_novel_detail_cover"
@@ -362,10 +363,10 @@
                                 android:layout_width="0dp"
                                 android:layout_height="match_parent"
                                 android:layout_marginEnd="4dp"
-                                android:onClick="@{() -> updateUserInterest.invoke()}"
                                 android:layout_weight="1"
                                 android:background="@drawable/bg_novel_detail_interest_selector"
-                                android:gravity="center">
+                                android:gravity="center"
+                                android:onClick="@{() -> updateUserInterest.invoke()}">
 
                                 <ImageView
                                     android:layout_width="16dp"
