+++ app/src/main/res/layout/fragment_novel_info.xml
@@ -109,6 +109,7 @@
                     android:layout_width="46dp"
                     android:layout_height="48dp"
                     android:layout_marginHorizontal="8dp"
+                    app:isVectorImage="@{true}"
                     android:onClick="@{() -> navigateToReadNovel.invoke(novelInfoViewModel.uiState.platforms.naverModel.platformUrl)}"
                     app:cornerRadius="@{24f}"
                     app:isVisible="@{novelInfoViewModel.uiState.platforms.naverModel.isVisible}"
@@ -119,6 +120,7 @@
                 <ImageView
                     android:id="@+id/iv_novel_info_kakao"
                     android:layout_width="46dp"
+                    app:isVectorImage="@{true}"
                     android:layout_height="48dp"
                     android:layout_marginHorizontal="8dp"
                     android:onClick="@{() -> navigateToReadNovel.invoke(novelInfoViewModel.uiState.platforms.kakaoModel.platformUrl)}"
